requirements.txt
Package                   Version
------------------------- --------------
absl-py                   2.1.0
aiohappyeyeballs          2.4.0
aiohttp                   3.10.5
aiosignal                 1.3.1
annotated-types           0.7.0
anyio                     4.4.0
argon2-cffi               23.1.0
argon2-cffi-bindings      21.2.0
arrow                     1.3.0
asttokens                 2.4.1
async-lru                 2.0.4
async-timeout             4.0.3
attrs                     24.2.0
babel                     2.16.0
backoff                   2.2.1
beautifulsoup4            4.12.3
bleach                    6.1.0
boto3                     1.35.12
botocore                  1.35.12
cachetools                5.5.0
causal-learn              0.1.3.8
cdt                       0.6.0
certifi                   2024.8.30
cffi                      1.17.1
charset-normalizer        3.3.2
clarabel                  0.9.0
click                     8.1.7
cloudpickle               3.1.0
comm                      0.2.2
contourpy                 1.3.0
cvxpy                     1.5.3
cycler                    0.12.1
Cython                    0.29.37
debugpy                   1.8.5
decorator                 5.1.1
defusedxml                0.7.1
dowhy                     0.0.0
econml                    0.15.1
ecos                      2.0.14
exceptiongroup            1.2.2
executing                 2.1.0
fastapi                   0.113.0
fastjsonschema            2.20.0
filelock                  3.15.4
fire                      0.6.0
fonttools                 4.53.1
fqdn                      1.5.1
frozenlist                1.4.1
fsspec                    2024.9.0
gcastle                   1.0.4rc1
google-auth               2.34.0
google-auth-oauthlib      1.2.1
GPUtil                    1.4.0
graphviz                  0.20.3
grpcio                    1.66.1
h11                       0.14.0
httpcore                  1.0.5
httpx                     0.27.2
idna                      3.8
ipykernel                 6.26.0
ipython                   8.17.2
ipywidgets                8.1.1
isoduration               20.11.0
jedi                      0.19.1
Jinja2                    3.1.4
jmespath                  1.0.1
joblib                    1.4.2
json5                     0.9.25
jsonpointer               3.0.0
jsonschema                4.23.0
jsonschema-specifications 2023.12.1
jupyter_client            8.6.2
jupyter_core              5.7.2
jupyter-events            0.10.0
jupyter-lsp               2.2.5
jupyter_server            2.14.2
jupyter_server_terminals  0.5.3
jupyterlab                4.2.0
jupyterlab_pygments       0.3.0
jupyterlab_server         2.27.3
jupyterlab_widgets        3.0.13
kiwisolver                1.4.7
lightgbm                  4.5.0
lightning                 2.4.0
lightning-cloud           0.5.70
lightning_sdk             0.1.15
lightning-utilities       0.11.7
lingam                    1.9.0
litdata                   0.2.19
litserve                  0.2.2
llvmlite                  0.43.0
Markdown                  3.7
markdown-it-py            3.0.0
MarkupSafe                2.1.5
matplotlib                3.8.2
matplotlib-inline         0.1.7
mdurl                     0.1.2
mistune                   3.0.2
mpmath                    1.3.0
multidict                 6.0.5
nbclient                  0.10.0
nbconvert                 7.16.4
nbformat                  5.10.4
nest-asyncio              1.6.0
networkx                  2.8.8
notebook_shim             0.2.4
numba                     0.60.0
numdifftools              0.9.41
numpy                     1.26.4
nvidia-cublas-cu12        12.1.3.1
nvidia-cuda-cupti-cu12    12.1.105
nvidia-cuda-nvrtc-cu12    12.1.105
nvidia-cuda-runtime-cu12  12.1.105
nvidia-cudnn-cu12         8.9.2.26
nvidia-cufft-cu12         11.0.2.54
nvidia-curand-cu12        10.3.2.106
nvidia-cusolver-cu12      11.4.5.107
nvidia-cusparse-cu12      12.1.0.106
nvidia-nccl-cu12          2.19.3
nvidia-nvjitlink-cu12     12.6.68
nvidia-nvtx-cu12          12.1.105
oauthlib                  3.2.2
osqp                      0.6.7.post3
overrides                 7.7.0
packaging                 24.1
pandas                    2.1.4
pandocfilters             1.5.1
parso                     0.8.4
patsy                     0.5.6
pexpect                   4.9.0
pillow                    10.4.0
pip                       24.2
platformdirs              4.2.2
progressbar2              4.5.0
prometheus_client         0.20.0
prompt_toolkit            3.0.47
protobuf                  4.23.4
psutil                    6.0.0
psy                       0.0.1
ptyprocess                0.7.0
pure_eval                 0.2.3
pyasn1                    0.6.0
pyasn1_modules            0.4.0
pycparser                 2.22
pydantic                  2.9.0
pydantic_core             2.23.2
pydot                     3.0.2
pygam                     0.9.1
Pygments                  2.18.0
PyJWT                     2.9.0
pyparsing                 3.1.4
python-dateutil           2.9.0.post0
python-json-logger        2.0.7
python-multipart          0.0.9
python-utils              3.9.0
pytorch-lightning         2.4.0
pytz                      2024.1
PyYAML                    6.0.2
pyzmq                     26.2.0
qdldl                     0.1.7.post4
referencing               0.35.1
requests                  2.32.3
requests-oauthlib         2.0.0
rfc3339-validator         0.1.4
rfc3986-validator         0.1.1
rich                      13.8.0
rpds-py                   0.20.0
rsa                       4.9
s3transfer                0.10.2
scikit-learn              1.3.2
scipy                     1.11.4
scs                       3.2.7
seaborn                   0.13.2
semopy                    2.3.11
Send2Trash                1.8.3
setuptools                72.1.0
shap                      0.43.0
simple-term-menu          1.6.4
six                       1.16.0
skrebate                  0.62
slicer                    0.0.7
sniffio                   1.3.1
soupsieve                 2.6
sparse                    0.15.4
stack-data                0.6.3
starlette                 0.38.4
statsmodels               0.14.4
sympy                     1.13.2
tensorboard               2.15.1
tensorboard-data-server   0.7.2
termcolor                 2.4.0
terminado                 0.18.1
threadpoolctl             3.5.0
tinycss2                  1.3.0
tomli                     2.0.1
torch                     2.2.1+cu121
torchmetrics              1.3.1
torchvision               0.17.1+cu121
tornado                   6.4.1
tqdm                      4.66.5
traitlets                 5.14.3
triton                    2.2.0
types-python-dateutil     2.9.0.20240821
typing_extensions         4.12.2
tzdata                    2024.1
uri-template              1.3.0
urllib3                   2.2.2
uvicorn                   0.30.6
wcwidth                   0.2.13
webcolors                 24.8.0
webencodings              0.5.1
websocket-client          1.8.0
Werkzeug                  3.0.4
wheel                     0.43.0
widgetsnbextension        4.0.13
yarl                      1.9.11
